feat(datetime-navigator): move aotw models to lib directory
