+++ src/app/components/lib/chip/chip.model.ts
@@ -1,0 +1,6 @@
+export interface Chip {
+  label: string;
+  icon?: string;
+  active: boolean;
+  disabled: boolean;
+}
