+++ src/app/components/lib/chip/index.ts
@@ -1,1 +1,2 @@
 export * from './chip.component';
+export * from './chip.model';
