+++ src/app/components/lib/stepper/index.ts
@@ -1,2 +1,3 @@
 export * from './stepper.component';
+export * from './stepper.model';
 export * from './stepper.service';
