+++ src/app/components/lib/stepper/stepper.model.ts
@@ -1,0 +1,4 @@
+export interface Step {
+	label: string;
+	disabled: boolean;
+}
