+++ src/app/components/lib/stepper/stepper.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Subject } from 'rxjs';
 
-import { Step } from '../../../models';
+import { Step } from './stepper.model';
 
 @Injectable({
   providedIn: 'root'
