+++ src/app/components/lib/tab-group/index.ts
@@ -1,1 +1,2 @@
 export * from './tab-group.component';
+export * from './tab-group.model';
