+++ src/app/components/lib/tab-group/tab-group.model.ts
@@ -1,0 +1,6 @@
+export interface Tab {
+  id: number;
+  name: string;
+  label: string;
+  disabled: boolean;
+}
